requirements.txt
requests>=2.31
python-dotenv>=1.0
rank-bm25>=0.2.2
sentence-transformers>=2.7
transformers>=4.41,<5
numpy>=1.26
google-genai>=1.0
groq>=0.18
jupyter>=1.0
ipykernel>=6.0
pandas>=2.0
